split Buffer obj file to Layout and Mesh

+++ DI/src/core/Time.cpp
@@ -3,8 +3,8 @@
 
 namespace DI{
 
-double CoreTime::time_since_start_programm = 0;
-double CoreTime::tic = 0;
+float CoreTime::time_since_start_programm = 0;
+float CoreTime::tic = 0;
 
 CoreTime::CoreTime(){
    start = std::chrono::steady_clock::now();
@@ -12,7 +12,7 @@ CoreTime::CoreTime(){
 
 CoreTime::~CoreTime(){
    end = std::chrono::steady_clock::now();
-   std::chrono::duration<double> elapsed_seconds = end - start;
+   std::chrono::duration<float> elapsed_seconds = end - start;
    tic = elapsed_seconds.count();
    time_since_start_programm += elapsed_seconds.count();
 }
